Use futures::executor::block_on in tests
This function had a conditional implementation in the async shim,
but we can use the futures version everywhere. The cost is making
more test functions explicitly async, but I think that is okay.

The benchmarks can now run for the tokio runtime as well, though
they succeed in completing on neither runtime for me. I have not
yet pinned down why, or if it is a regression in this work or not.

Added test and bench invocations in the justfile for both runtimes.